Moving new failure test to smaller file
